requirements/development.txt
-r base.txt

django-debug-toolbar==4.1.0
ipython==8.12.0
black==23.7.0
flake8==6.0.0
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0
faker==19.2.0
